refactor: date formatter modulation
